Revamp mode management
Move everything dealing with the bindable/setable interfaces into
package mode, which is now compliant with bindable and setable.

Users should dot-import the entire package, which provides terse
functions to support Bind, stateful Set (see below), and SetMode.

Add stateful decorator to allow opt-in mode management for calls
to Set(), where the value of any call is stored but only applied to the
device when in the specified mode. Upon state change, the last value is
automatically applied.