apps/selah: implement (solo, mute, rec)
It turns out the best way to do this is to handle Bind/Send for each
device independently, track source-of-truth state (i.e. state
according to reaper) locally within TrackData, and implement whatever
conversion and mode logic is necessary per control.

Importantly, the feedback to xtouch is display-only, meaning there is no
feedback loop from xtouch that requires deduplication.

Feedback from reaper will always match the last-set value. Even if it
didn't, this configuration fully makes xtouch the slave of reaper for
these controls, so there is nothing to worry about.

A corrolary of this work is that xtouch does not need any state tracking
per-button; all such state tracking is better hoisted up a layer. This
will cause us to greatly simplify the implementation of Button in xtouch
in an upcoming commit.