osc: return unregister function from all Bind()
If we know we don't need to handle a route anymore, or its behavior
needs to change, calling unregister lets us stop handling it.

This change affects both the base osc implementation and the
implementation for reaper.

Since Bind() returned nothing, current users are not impacted since they
weren't checking for a return argument anyway. The only exception is in
app/selah/layers/tracks.go, since it was calling AddMsgHandler directly
and checking for an error (which could never return non-nil anyway).